@@ -6,7 +6,7 @@ version: 1.9.7
 homepage: https://github.com/FilledStacks/stacked
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
