@@ -10,35 +10,35 @@ import 'package:stacked/src/code_generation/router/router_utils.dart';
 
 import 'uri_extension.dart';
 
-RectTween _createRectTween(Rect begin, Rect end) {
+RectTween _createRectTween(Rect? begin, Rect? end) {
   return MaterialRectArcTween(begin: begin, end: end);
 }
 
 extension BuildContextX on BuildContext {
-  ExtendedNavigatorState get navigator =>
+  ExtendedNavigatorState? get navigator =>
       ExtendedNavigator.of(this, nullOk: true);
 
-  ExtendedNavigatorState get rootNavigator =>
+  ExtendedNavigatorState? get rootNavigator =>
       ExtendedNavigator.of(this, rootRouter: true, nullOk: true);
 }
 
-typedef OnNavigationRejected = void Function(RouteGuard guard);
+typedef OnNavigationRejected = void Function(RouteGuard? guard);
 
-class ExtendedNavigator<T extends RouterBase> extends StatefulWidget {
+class ExtendedNavigator<T extends RouterBase?> extends StatefulWidget {
   static TransitionBuilder builder<T extends RouterBase>({
-    GlobalKey<NavigatorState> navigatorKey,
+    GlobalKey<NavigatorState>? navigatorKey,
     List<NavigatorObserver> observers = const [],
-    RouteFactory onUnknownRoute,
-    String initialRoute,
-    Object initialRouteArgs,
-    @required T router,
+    RouteFactory? onUnknownRoute,
+    String? initialRoute,
+    Object? initialRouteArgs,
+    required T router,
     List<RouteGuard> guards = const [],
-    String name,
-    TransitionBuilder builder,
+    String? name,
+    TransitionBuilder? builder,
   }) =>
       (context, nav) {
         var extendedNav = ExtendedNavigator<T>(
-          key: GlobalObjectKey(nav.key),
+          key: GlobalObjectKey(nav!.key!),
           navigatorKey: navigatorKey,
           observers: observers,
           onUnknownRoute: onUnknownRoute,
@@ -63,17 +63,17 @@ class ExtendedNavigator<T extends RouterBase> extends StatefulWidget {
     this.initialRouteArgs,
     this.navigatorKey,
     this.observers = const <NavigatorObserver>[],
-    Key key,
+    Key? key,
   }) : super(key: key);
 
-  final GlobalKey<NavigatorState> navigatorKey;
+  final GlobalKey<NavigatorState>? navigatorKey;
   final List<NavigatorObserver> observers;
-  final RouteFactory onUnknownRoute;
-  final String initialRoute;
-  final Object initialRouteArgs;
-  final T router;
+  final RouteFactory? onUnknownRoute;
+  final String? initialRoute;
+  final Object? initialRouteArgs;
+  final T? router;
   final List<RouteGuard> guards;
-  final String name;
+  final String? name;
 
   @override
   ExtendedNavigatorState createState() => ExtendedNavigatorState<T>();
@@ -84,13 +84,13 @@ class ExtendedNavigator<T extends RouterBase> extends StatefulWidget {
         ),
       );
 
-  static List<Route<dynamic>> _generateInitialRoutes(
+  static List<Route<dynamic>?> _generateInitialRoutes(
     ExtendedNavigatorState router,
     Uri initialUri,
-    Object initialRouteArgs,
+    Object? initialRouteArgs,
   ) {
-    var initialRoutes = <Route>[];
-    var navigator = router._navigator;
+    var initialRoutes = <Route?>[];
+    var navigator = router._navigator!;
 
     var initialRouteName = initialUri.path;
     if (!kIsWeb &&
@@ -98,7 +98,7 @@ class ExtendedNavigator<T extends RouterBase> extends StatefulWidget {
         initialRouteName.length > 1) {
       var root = initialUri.replace(path: "/").normalizedPath;
       var rootRoute =
-          navigator.widget.onGenerateRoute(RouteSettings(name: root));
+          navigator.widget.onGenerateRoute!(RouteSettings(name: root));
       if (rootRoute != null) {
         if ((rootRoute.settings as RouteData).routeMatch.hasGuards) {
           initialRoutes.add(_placeHolderRoute);
@@ -111,7 +111,7 @@ class ExtendedNavigator<T extends RouterBase> extends StatefulWidget {
 
     var settings = RouteSettings(
         name: initialUri.normalizedPath, arguments: initialRouteArgs);
-    var route = navigator.widget.onGenerateRoute(settings);
+    var route = navigator.widget.onGenerateRoute!(settings);
 
     if (route != null) {
       var data = route.settings as RouteData;
@@ -125,7 +125,7 @@ class ExtendedNavigator<T extends RouterBase> extends StatefulWidget {
         initialRoutes.add(route);
       }
     } else {
-      var unknownRoute = navigator.widget.onUnknownRoute(settings);
+      var unknownRoute = navigator.widget.onUnknownRoute!(settings);
       initialRoutes.add(unknownRoute);
     }
 
@@ -136,20 +136,20 @@ class ExtendedNavigator<T extends RouterBase> extends StatefulWidget {
   @Deprecated("use ExtendedNavigator.builder instead")
   ExtendedNavigator<T> call(_, navigator) => this;
 
-  static ExtendedNavigatorState named(String name) {
+  static ExtendedNavigatorState? named(String name) {
     assert(name != null);
     return _NavigatorsContainer._instance.get(name);
   }
 
-  static ExtendedNavigatorState get root =>
+  static ExtendedNavigatorState? get root =>
       _NavigatorsContainer._instance.getRootRouter();
 
-  static ExtendedNavigatorState of(
+  static ExtendedNavigatorState? of(
     BuildContext context, {
     bool rootRouter = false,
     bool nullOk = false,
   }) {
-    final ExtendedNavigatorState router = rootRouter
+    final ExtendedNavigatorState? router = rootRouter
         ? context.findRootAncestorStateOfType<ExtendedNavigatorState>()
         : context.findAncestorStateOfType<ExtendedNavigatorState>();
     assert(() {
@@ -165,37 +165,37 @@ class ExtendedNavigator<T extends RouterBase> extends StatefulWidget {
   }
 }
 
-class ExtendedNavigatorState<T extends RouterBase>
+class ExtendedNavigatorState<T extends RouterBase?>
     extends State<ExtendedNavigator<T>> with WidgetsBindingObserver {
   final _registeredGuards = <Type, RouteGuard>{};
-  T router;
+  T? router;
   List<Route> _guardedInitialRoutes = [];
 
-  ExtendedNavigatorState get parent => _parent;
+  ExtendedNavigatorState? get parent => _parent;
 
-  ExtendedNavigatorState _parent;
-  final children = <ExtendedNavigatorState>[];
-  GlobalKey<NavigatorState> _navigatorKey;
+  ExtendedNavigatorState? _parent;
+  final List<ExtendedNavigatorState<RouterBase?>> children = <ExtendedNavigatorState>[];
+  GlobalKey<NavigatorState>? _navigatorKey;
 
-  NavigatorState get _navigator => _navigatorKey.currentState;
+  NavigatorState? get _navigator => _navigatorKey!.currentState;
 
   bool get isRoot => parent == null;
 
-  WillPopCallback _willPopCallback;
+  WillPopCallback? _willPopCallback;
 
   ExtendedNavigatorState get root =>
       context.findRootAncestorStateOfType<ExtendedNavigatorState>() ?? this;
 
-  HeroController _heroController;
+  HeroController? _heroController;
 
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
     _navigatorKey = widget.navigatorKey ?? GlobalKey<NavigatorState>();
     _heroController = HeroController(createRectTween: _createRectTween);
     if (_guardedInitialRoutes.isNotEmpty) {
-      SchedulerBinding.instance.addPostFrameCallback((_) {
+      SchedulerBinding.instance!.addPostFrameCallback((_) {
         _pushAllGuarded(_guardedInitialRoutes);
       });
     }
@@ -206,10 +206,10 @@ class ExtendedNavigatorState<T extends RouterBase>
       var data = (route.settings as RouteData);
       if (await _canNavigate(data.template)) {
         if (data.template == Navigator.defaultRouteName) {
-          _navigator.pushAndRemoveUntil(
+          _navigator!.pushAndRemoveUntil(
               route, RouteData.withPath(data.template));
         } else {
-          _navigator.push(route);
+          _navigator!.push(route);
         }
       } else {
         break;
@@ -222,7 +222,7 @@ class ExtendedNavigatorState<T extends RouterBase>
   Future<bool> didPopRoute() async {
     assert(mounted);
     if (_navigator == null) return false;
-    return await _navigator.maybePop();
+    return await _navigator!.maybePop();
   }
 
   @override
@@ -241,7 +241,7 @@ class ExtendedNavigatorState<T extends RouterBase>
     var basePath;
     var parentData = ParentRouteData.of(context);
     if (router == null && parentData != null) {
-      router = parentData.router;
+      router = parentData.router as T?;
       basePath = parentData.name;
       if (parentData.initialRouteArgs != null) {
         initialRouteArgs = parentData.initialRouteArgs;
@@ -251,20 +251,20 @@ class ExtendedNavigatorState<T extends RouterBase>
     assert(router != null);
     return Navigator(
       key: _navigatorKey,
-      initialRoute: WidgetsBinding.instance.window.defaultRouteName !=
+      initialRoute: WidgetsBinding.instance!.window.defaultRouteName !=
               Navigator.defaultRouteName
-          ? WidgetsBinding.instance.window.defaultRouteName
-          : initial ?? WidgetsBinding.instance.window.defaultRouteName,
-      observers: List.from(widget.observers)..add(_heroController),
+          ? WidgetsBinding.instance!.window.defaultRouteName
+          : initial ?? WidgetsBinding.instance!.window.defaultRouteName,
+      observers: List.from(widget.observers)..add(_heroController!),
       onGenerateRoute: (RouteSettings settings) {
-        return router.onGenerateRoute(settings, basePath);
+        return router!.onGenerateRoute(settings, basePath);
       },
       onUnknownRoute: widget.onUnknownRoute ?? defaultUnknownRoutePage,
       onGenerateInitialRoutes: (NavigatorState navigator, String initialRoute) {
         var initialUri;
         if (parentData != null) {
-          if (parentData.initialRoute.hasEmptyPath) {
-            initialUri = parentData.initialRoute.replace(path: initialRoute);
+          if (parentData.initialRoute!.hasEmptyPath) {
+            initialUri = parentData.initialRoute!.replace(path: initialRoute);
           } else {
             initialUri = parentData.initialRoute;
           }
@@ -275,8 +275,8 @@ class ExtendedNavigatorState<T extends RouterBase>
           this,
           initialUri,
           initialRouteArgs,
-        );
-      },
+        ) as List<Route<dynamic>>;
+      } as List<Route<dynamic>> Function(NavigatorState, String),
     );
   }
 
@@ -286,12 +286,12 @@ class ExtendedNavigatorState<T extends RouterBase>
 
     _parent = context.findAncestorStateOfType<ExtendedNavigatorState>();
     if (_parent != null) {
-      _parent.children.add(this);
-      if (_parent._registeredGuards != null) {
-        _registeredGuards.addAll(_parent._registeredGuards);
+      _parent!.children.add(this);
+      if (_parent!._registeredGuards != null) {
+        _registeredGuards.addAll(_parent!._registeredGuards);
       }
     }
-    widget.guards?.forEach(_registerGuard);
+    widget.guards.forEach(_registerGuard);
 
     if (routerName != null) {
       _NavigatorsContainer._instance.register(this, name: routerName);
@@ -299,32 +299,31 @@ class ExtendedNavigatorState<T extends RouterBase>
     final modelRoute = ModalRoute.of(context);
     if (modelRoute != null) {
       modelRoute.addScopedWillPopCallback(_willPopCallback = () async {
-        return !await _navigator.maybePop();
+        return !await _navigator!.maybePop();
       });
     }
   }
 
   void _registerGuard(RouteGuard guard) {
-    assert(guard != null);
     _registeredGuards[guard.runtimeType] = guard;
   }
 
   @optionalTypeArgs
-  Future<T> push<T extends Object>(String routeName,
-      {Object arguments,
-      Map<String, String> queryParams,
-      OnNavigationRejected onReject}) async {
+  Future<T?>? push<T extends Object>(String routeName,
+      {Object? arguments,
+      Map<String, String>? queryParams,
+      OnNavigationRejected? onReject}) async {
     return await _canNavigate(
       routeName,
       arguments: arguments,
       onReject: onReject,
     )
-        ? _navigator.pushNamed<T>(_buildPath(routeName, queryParams),
+        ? _navigator!.pushNamed<T>(_buildPath(routeName, queryParams),
             arguments: arguments)
         : null;
   }
 
-  String _buildPath(String routeName, Map<String, String> queryParams) {
+  String _buildPath(String routeName, Map<String, String>? queryParams) {
     var uri = Uri.parse(routeName);
     var params = <String, String>{};
     if (queryParams != null) {
@@ -335,16 +334,16 @@ class ExtendedNavigatorState<T extends RouterBase>
   }
 
   @optionalTypeArgs
-  Future<T> replace<T extends Object, TO extends Object>(
+  Future<T?>? replace<T extends Object, TO extends Object>(
     String routeName, {
-    TO result,
-    Object arguments,
-    Map<String, String> queryParams,
-    OnNavigationRejected onReject,
+    TO? result,
+    Object? arguments,
+    Map<String, String>? queryParams,
+    OnNavigationRejected? onReject,
   }) async {
     return await _canNavigate(routeName,
             arguments: arguments, onReject: onReject)
-        ? _navigator.pushReplacementNamed<T, TO>(
+        ? _navigator!.pushReplacementNamed<T, TO>(
             _buildPath(routeName, queryParams),
             arguments: arguments,
             result: result)
@@ -352,16 +351,16 @@ class ExtendedNavigatorState<T extends RouterBase>
   }
 
   @optionalTypeArgs
-  Future<T> pushAndRemoveUntil<T extends Object>(
+  Future<T?>? pushAndRemoveUntil<T extends Object>(
     String newRouteName,
     RoutePredicate predicate, {
-    Object arguments,
-    Map<String, String> queryParams,
-    OnNavigationRejected onReject,
+    Object? arguments,
+    Map<String, String>? queryParams,
+    OnNavigationRejected? onReject,
   }) async {
     return await _canNavigate(newRouteName,
             arguments: arguments, onReject: onReject)
-        ? _navigator.pushNamedAndRemoveUntil<T>(
+        ? _navigator!.pushNamedAndRemoveUntil<T>(
             _buildPath(newRouteName, queryParams),
             predicate,
             arguments: arguments,
@@ -370,12 +369,12 @@ class ExtendedNavigatorState<T extends RouterBase>
   }
 
   @optionalTypeArgs
-  Future<T> pushAndRemoveUntilPath<T extends Object>(
+  Future<T?>? pushAndRemoveUntilPath<T extends Object>(
     String newRouteName,
     String anchorPath, {
-    Object arguments,
-    Map<String, String> queryParams,
-    OnNavigationRejected onReject,
+    Object? arguments,
+    Map<String, String>? queryParams,
+    OnNavigationRejected? onReject,
   }) {
     return pushAndRemoveUntil(
       newRouteName,
@@ -387,12 +386,12 @@ class ExtendedNavigatorState<T extends RouterBase>
   }
 
   @optionalTypeArgs
-  Future<T> popAndPush<T extends Object, TO extends Object>(
+  Future<T?>? popAndPush<T extends Object, TO extends Object>(
     String routeName, {
-    TO result,
-    Object arguments,
-    Map<String, String> queryParams,
-    OnNavigationRejected onReject,
+    TO? result,
+    Object? arguments,
+    Map<String, String>? queryParams,
+    OnNavigationRejected? onReject,
   }) {
     pop<TO>(result);
     return push<T>(routeName,
@@ -404,7 +403,7 @@ class ExtendedNavigatorState<T extends RouterBase>
   }
 
   void popUntil(RoutePredicate predicate) {
-    _navigator.popUntil(predicate);
+    _navigator!.popUntil(predicate);
   }
 
   void popUntilRoot() {
@@ -412,29 +411,29 @@ class ExtendedNavigatorState<T extends RouterBase>
   }
 
   @optionalTypeArgs
-  void pop<T extends Object>([T result]) {
-    _navigator.pop<T>(result);
+  void pop<T extends Object>([T? result]) {
+    _navigator!.pop<T>(result);
   }
 
-  Future<bool> maybePop<T extends Object>([T result]) async {
-    return _navigator.maybePop(result);
+  Future<bool> maybePop<T extends Object>([T? result]) async {
+    return _navigator!.maybePop(result);
   }
 
   bool canPop() {
-    return _navigator.canPop();
+    return _navigator!.canPop();
   }
 
   Future<bool> canNavigate(String routeName) => _canNavigate(routeName);
 
   Future<bool> _canNavigate(String routeName,
-      {Object arguments, OnNavigationRejected onReject}) async {
-    var match = router.findMatch(RouteSettings(name: routeName));
+      {Object? arguments, OnNavigationRejected? onReject}) async {
+    var match = router!.findMatch(RouteSettings(name: routeName));
     if (match == null || !match.hasGuards) {
       return true;
     }
 
-    for (Type guardType in match.routeDef.guards) {
-      if (!await _getGuard(guardType).canNavigate(this, routeName, arguments)) {
+    for (Type guardType in match.routeDef.guards!) {
+      if (!await _getGuard(guardType)!.canNavigate(this, routeName, arguments)) {
         if (onReject != null) {
           onReject(_getGuard(guardType));
         }
@@ -444,7 +443,7 @@ class ExtendedNavigatorState<T extends RouterBase>
     return true;
   }
 
-  RouteGuard _getGuard(Type guardType) {
+  RouteGuard? _getGuard(Type guardType) {
     if (_registeredGuards[guardType] == null) {
       throw ('$guardType is not registered!'
           '\nYou have to add your guards to ExtendedNavigator widget');
@@ -455,7 +454,7 @@ class ExtendedNavigatorState<T extends RouterBase>
   @override
   void deactivate() {
     if (_willPopCallback != null) {
-      ModalRoute.of(context).removeScopedWillPopCallback(_willPopCallback);
+      ModalRoute.of(context)!.removeScopedWillPopCallback(_willPopCallback!);
     }
     super.deactivate();
   }
@@ -463,9 +462,9 @@ class ExtendedNavigatorState<T extends RouterBase>
   @override
   void dispose() {
     if (_parent != null) {
-      _parent.children.remove(this);
+      _parent!.children.remove(this);
     }
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
 
     if (routerName != null) {
       _NavigatorsContainer._instance.remove(routerName);
@@ -473,7 +472,7 @@ class ExtendedNavigatorState<T extends RouterBase>
     super.dispose();
   }
 
-  String get routerName {
+  String? get routerName {
     if (widget.name != null) {
       return widget.name;
     } else {
@@ -489,26 +488,26 @@ class ExtendedNavigatorState<T extends RouterBase>
 class _NavigatorsContainer {
   // ensure we only have one instance of the container
   static final _NavigatorsContainer _instance = _NavigatorsContainer._();
-  final Map<String, ExtendedNavigatorState> _routers = {};
+  final Map<String?, ExtendedNavigatorState> _routers = {};
 
   _NavigatorsContainer._();
 
-  void register(ExtendedNavigatorState router, {String name}) {
+  void register(ExtendedNavigatorState router, {String? name}) {
     _routers[name] = router;
   }
 
-  ExtendedNavigatorState getRootRouter() {
+  ExtendedNavigatorState? getRootRouter() {
     if (_routers.isNotEmpty) {
       return _routers.values.first.root;
     }
     return null;
   }
 
-  ExtendedNavigatorState remove(String name) {
+  ExtendedNavigatorState? remove(String? name) {
     return _routers.remove(name);
   }
 
-  ExtendedNavigatorState get<T extends RouterBase>([String name]) {
+  ExtendedNavigatorState? get<T extends RouterBase>([String? name]) {
     return _routers[name ?? T.toString()];
   }
 }
