@@ -6,7 +6,7 @@ import 'package:stacked/src/state_management/view_model_builder.dart';
 /// Default [reactive] value is true. Can be overriden and set to false
 abstract class ViewModelBuilderWidget<T extends ChangeNotifier>
     extends StatelessWidget {
-  const ViewModelBuilderWidget({Key key}) : super(key: key);
+  const ViewModelBuilderWidget({Key? key}) : super(key: key);
 
   /// A function that builds the UI to be shown from the ViewModel - Required
   ///
@@ -56,7 +56,7 @@ abstract class ViewModelBuilderWidget<T extends ChangeNotifier>
   ///
   /// When [reactive] is set to false the builder is used as the static child
   /// and is only ever built once.
-  Widget staticChildBuilder(BuildContext context) => null;
+  Widget? staticChildBuilder(BuildContext context) => null;
 
   @override
   Widget build(BuildContext context) {
@@ -65,7 +65,7 @@ abstract class ViewModelBuilderWidget<T extends ChangeNotifier>
         builder: builder,
         viewModelBuilder: () => viewModelBuilder(context),
         staticChild:
-            staticChildBuilder != null ? staticChildBuilder(context) : null,
+            staticChildBuilder(context) ,
         onModelReady: onViewModelReady,
         disposeViewModel: disposeViewModel,
         createNewModelOnInsert: createNewModelOnInsert,
