@@ -6,7 +6,7 @@ import 'package:provider/provider.dart';
 abstract class ViewModelWidget<T> extends Widget {
   final bool reactive;
 
-  const ViewModelWidget({Key key, this.reactive = true}) : super(key: key);
+  const ViewModelWidget({Key? key, this.reactive = true}) : super(key: key);
 
   @protected
   Widget build(BuildContext context, T viewModel);
@@ -19,7 +19,7 @@ class _DataProviderElement<T> extends ComponentElement {
   _DataProviderElement(ViewModelWidget widget) : super(widget);
 
   @override
-  ViewModelWidget get widget => super.widget;
+  ViewModelWidget get widget => super.widget as ViewModelWidget<dynamic>;
 
   @override
   Widget build() =>
