@@ -5,6 +5,6 @@ abstract class RouteGuard {
   Future<bool> canNavigate(
     ExtendedNavigatorState navigator,
     String routeName,
-    Object arguments,
+    Object? arguments,
   );
 }
