@@ -2,5 +2,5 @@ extension UriX on Uri {
   String get normalizedPath =>
       hasQueryParams || hasFragment ? toString() : path;
 
-  bool get hasQueryParams => queryParameters?.isNotEmpty == true;
+  bool get hasQueryParams => queryParameters.isNotEmpty == true;
 }
